feat: simplify plugin tests; fix plugin resetting
