+++ test/plugins/test_manager.py
@@ -3,47 +3,24 @@
 import pytest
 
 from mellea.plugins.base import MelleaBasePayload
-from mellea.plugins.manager import (
-    ensure_plugin_manager,
-    has_plugins,
-    invoke_hook,
-    shutdown_plugins,
-)
+from mellea.plugins.manager import ensure_plugin_manager, has_plugins, invoke_hook
 from mellea.plugins.types import HookType, PluginMode
 
 # These tests require the contextforge plugin framework
 pytest.importorskip("cpex.framework")
 
 
-@pytest.fixture(autouse=True)
-async def cleanup_plugins():
-    """Ensure plugin manager is shut down after each test."""
-    yield
-    await shutdown_plugins()
-
-
 class TestNoOpGuards:
     @pytest.mark.asyncio
-    async def test_invoke_hook_noop_when_no_plugins(self, request):
+    async def test_invoke_hook_noop_when_no_plugins(self):
         """When no plugins are registered, invoke_hook returns (None, original_payload)."""
-        plugins_disabled = request.config.getoption(
-            "--disable-default-mellea-plugins", default=False
-        )
-        if not plugins_disabled:
-            pytest.skip("must pass --disable-default-mellea-plugins for this test")
         payload = MelleaBasePayload(request_id="test-123")
         result, returned_payload = await invoke_hook(HookType.SESSION_PRE_INIT, payload)
         assert result is None
         assert returned_payload is payload
 
-    def test_has_plugins_false_by_default(self, request):
+    def test_has_plugins_false_by_default(self):
         """has_plugins() returns False when no plugins have been registered."""
-        plugins_disabled = request.config.getoption(
-            "--disable-default-mellea-plugins", default=False
-        )
-        if not plugins_disabled:
-            pytest.skip("must pass --disable-default-mellea-plugins for this test")
-        # After shutdown, plugins should not be enabled
         assert not has_plugins()
