+++ test/plugins/test_unregister.py
@@ -8,14 +8,8 @@ pytest.importorskip("cpex.framework")
 
 from mellea.plugins import Plugin, PluginSet, hook, register, unregister
 from mellea.plugins.hooks.session import SessionPreInitPayload
-from mellea.plugins.manager import invoke_hook, shutdown_plugins
+from mellea.plugins.manager import invoke_hook
 from mellea.plugins.types import HookType
-
-
-@pytest.fixture(autouse=True)
-async def reset_plugins():
-    yield
-    await shutdown_plugins()
 
 
 def _payload() -> SessionPreInitPayload:
